requirements.txt
Django>=4.2.0
requests
django-ratelimit
celery
redis
drf-yasg
djangorestframework
gunicorn
whitenoise
python-decouple